+++ src/main/java/com/workwave/entity/board/Board.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 @Builder
 public class Board {
 
-    private long boardId;
+    private int boardId;
     private String userId;
     private String boardTitle;
     private String boardContent;
